Moved BackupContainerLocalDirectory into its own files
