+++ fdbclient/BackupContainerFileSystem.actor.h
@@ -33,7 +33,6 @@
 #include "flow/Platform.actor.h"
 #include "fdbclient/AsyncTaskThread.h"
 #include "fdbclient/BackupContainer.h"
-#include "fdbclient/BackupContainerFileSystem.actor.h"
 #include "fdbclient/BackupAgent.actor.h"
 #include "fdbclient/FDBTypes.h"
 #include "fdbclient/JsonBuilder.h"
@@ -279,230 +278,6 @@ public:
 	static constexpr Version NON_PARTITIONED_MUTATION_LOG = 0;
 	static constexpr Version PARTITIONED_MUTATION_LOG = 1;
 	VersionProperty logType();
-};
-
-class BackupContainerLocalDirectory : public BackupContainerFileSystem,
-                                      ReferenceCounted<BackupContainerLocalDirectory> {
-public:
-	void addref() final { return ReferenceCounted<BackupContainerLocalDirectory>::addref(); }
-	void delref() final { return ReferenceCounted<BackupContainerLocalDirectory>::delref(); }
-
-	static std::string getURLFormat() { return "file://</path/to/base/dir/>"; }
-
-	BackupContainerLocalDirectory(std::string url) {
-		std::string path;
-		if (url.find("file://") != 0) {
-			TraceEvent(SevWarn, "BackupContainerLocalDirectory")
-			    .detail("Description", "Invalid URL for BackupContainerLocalDirectory")
-			    .detail("URL", url);
-		}
-
-		path = url.substr(7);
-		// Remove trailing slashes on path
-		path.erase(path.find_last_not_of("\\/") + 1);
-
-		std::string absolutePath = abspath(path);
-
-		if (!g_network->isSimulated() && path != absolutePath) {
-			TraceEvent(SevWarn, "BackupContainerLocalDirectory")
-			    .detail("Description", "Backup path must be absolute (e.g. file:///some/path)")
-			    .detail("URL", url)
-			    .detail("Path", path)
-			    .detail("AbsolutePath", absolutePath);
-			// throw io_error();
-			IBackupContainer::lastOpenError =
-			    format("Backup path '%s' must be the absolute path '%s'", path.c_str(), absolutePath.c_str());
-			throw backup_invalid_url();
-		}
-
-		// Finalized path written to will be will be <path>/backup-<uid>
-		m_path = path;
-	}
-
-	static Future<std::vector<std::string>> listURLs(std::string url) {
-		std::string path;
-		if (url.find("file://") != 0) {
-			TraceEvent(SevWarn, "BackupContainerLocalDirectory")
-			    .detail("Description", "Invalid URL for BackupContainerLocalDirectory")
-			    .detail("URL", url);
-		}
-
-		path = url.substr(7);
-		// Remove trailing slashes on path
-		path.erase(path.find_last_not_of("\\/") + 1);
-
-		if (!g_network->isSimulated() && path != abspath(path)) {
-			TraceEvent(SevWarn, "BackupContainerLocalDirectory")
-			    .detail("Description", "Backup path must be absolute (e.g. file:///some/path)")
-			    .detail("URL", url)
-			    .detail("Path", path);
-			throw io_error();
-		}
-		std::vector<std::string> dirs = platform::listDirectories(path);
-		std::vector<std::string> results;
-
-		for (auto& r : dirs) {
-			if (r == "." || r == "..") continue;
-			results.push_back(std::string("file://") + joinPath(path, r));
-		}
-
-		return results;
-	}
-
-	Future<Void> create() final {
-		// Nothing should be done here because create() can be called by any process working with the container URL,
-		// such as fdbbackup. Since "local directory" containers are by definition local to the machine they are
-		// accessed from, the container's creation (in this case the creation of a directory) must be ensured prior to
-		// every file creation, which is done in openFile(). Creating the directory here will result in unnecessary
-		// directories being created on machines that run fdbbackup but not agents.
-		return Void();
-	}
-
-	// The container exists if the folder it resides in exists
-	Future<bool> exists() final { return directoryExists(m_path); }
-
-	Future<Reference<IAsyncFile>> readFile(std::string path) final {
-		int flags = IAsyncFile::OPEN_NO_AIO | IAsyncFile::OPEN_READONLY | IAsyncFile::OPEN_UNCACHED;
-		// Simulation does not properly handle opening the same file from multiple machines using a shared filesystem,
-		// so create a symbolic link to make each file opening appear to be unique.  This could also work in production
-		// but only if the source directory is writeable which shouldn't be required for a restore.
-		std::string fullPath = joinPath(m_path, path);
-#ifndef _WIN32
-		if (g_network->isSimulated()) {
-			if (!fileExists(fullPath)) {
-				throw file_not_found();
-			}
-
-			if (g_simulator.getCurrentProcess()->uid == UID()) {
-				TraceEvent(SevError, "BackupContainerReadFileOnUnsetProcessID");
-			}
-			std::string uniquePath = fullPath + "." + g_simulator.getCurrentProcess()->uid.toString() + ".lnk";
-			unlink(uniquePath.c_str());
-			ASSERT(symlink(basename(path).c_str(), uniquePath.c_str()) == 0);
-			fullPath = uniquePath;
-		}
-// Opening cached mode forces read/write mode at a lower level, overriding the readonly request.  So cached mode
-// can't be used because backup files are read-only.  Cached mode can only help during restore task retries handled
-// by the same process that failed the first task execution anyway, which is a very rare case.
-#endif
-		Future<Reference<IAsyncFile>> f = IAsyncFileSystem::filesystem()->open(fullPath, flags, 0644);
-
-		if (g_network->isSimulated()) {
-			int blockSize = 0;
-			// Extract block size from the filename, if present
-			size_t lastComma = path.find_last_of(',');
-			if (lastComma != path.npos) {
-				blockSize = atoi(path.substr(lastComma + 1).c_str());
-			}
-			if (blockSize <= 0) {
-				blockSize = deterministicRandom()->randomInt(1e4, 1e6);
-			}
-			if (deterministicRandom()->random01() < .01) {
-				blockSize /= deterministicRandom()->randomInt(1, 3);
-			}
-			ASSERT(blockSize > 0);
-
-			return map(f, [=](Reference<IAsyncFile> fr) {
-				int readAhead = deterministicRandom()->randomInt(0, 3);
-				int reads = deterministicRandom()->randomInt(1, 3);
-				int cacheSize = deterministicRandom()->randomInt(0, 3);
-				return Reference<IAsyncFile>(new AsyncFileReadAheadCache(fr, blockSize, readAhead, reads, cacheSize));
-			});
-		}
-
-		return f;
-	}
-
-	class BackupFile : public IBackupFile, ReferenceCounted<BackupFile> {
-	public:
-		BackupFile(std::string fileName, Reference<IAsyncFile> file, std::string finalFullPath)
-		  : IBackupFile(fileName), m_file(file), m_finalFullPath(finalFullPath) {}
-
-		Future<Void> append(const void* data, int len) {
-			Future<Void> r = m_file->write(data, len, m_offset);
-			m_offset += len;
-			return r;
-		}
-
-		ACTOR static Future<Void> finish_impl(Reference<BackupFile> f) {
-			wait(f->m_file->truncate(f->size())); // Some IAsyncFile implementations extend in whole block sizes.
-			wait(f->m_file->sync());
-			std::string name = f->m_file->getFilename();
-			f->m_file.clear();
-			renameFile(name, f->m_finalFullPath);
-			return Void();
-		}
-
-		Future<Void> finish() { return finish_impl(Reference<BackupFile>::addRef(this)); }
-
-		void addref() override { return ReferenceCounted<BackupFile>::addref(); }
-		void delref() override { return ReferenceCounted<BackupFile>::delref(); }
-
-	private:
-		Reference<IAsyncFile> m_file;
-		std::string m_finalFullPath;
-	};
-
-	Future<Reference<IBackupFile>> writeFile(const std::string& path) final {
-		int flags = IAsyncFile::OPEN_NO_AIO | IAsyncFile::OPEN_CREATE | IAsyncFile::OPEN_ATOMIC_WRITE_AND_CREATE |
-		            IAsyncFile::OPEN_READWRITE;
-		std::string fullPath = joinPath(m_path, path);
-		platform::createDirectory(parentDirectory(fullPath));
-		std::string temp = fullPath + "." + deterministicRandom()->randomUniqueID().toString() + ".temp";
-		Future<Reference<IAsyncFile>> f = IAsyncFileSystem::filesystem()->open(temp, flags, 0644);
-		return map(f,
-		           [=](Reference<IAsyncFile> f) { return Reference<IBackupFile>(new BackupFile(path, f, fullPath)); });
-	}
-
-	Future<Void> deleteFile(std::string path) final {
-		::deleteFile(joinPath(m_path, path));
-		return Void();
-	}
-
-	ACTOR static Future<FilesAndSizesT> listFiles_impl(std::string path, std::string m_path) {
-		state std::vector<std::string> files;
-		wait(platform::findFilesRecursivelyAsync(joinPath(m_path, path), &files));
-
-		FilesAndSizesT results;
-
-		// Remove .lnk files from results, they are a side effect of a backup that was *read* during simulation.  See
-		// openFile() above for more info on why they are created.
-		if (g_network->isSimulated())
-			files.erase(
-			    std::remove_if(files.begin(), files.end(),
-			                   [](std::string const& f) { return StringRef(f).endsWith(LiteralStringRef(".lnk")); }),
-			    files.end());
-
-		for (auto& f : files) {
-			// Hide .part or .temp files.
-			StringRef s(f);
-			if (!s.endsWith(LiteralStringRef(".part")) && !s.endsWith(LiteralStringRef(".temp")))
-				results.push_back({ f.substr(m_path.size() + 1), ::fileSize(f) });
-		}
-
-		return results;
-	}
-
-	Future<FilesAndSizesT> listFiles(std::string path, std::function<bool(std::string const&)>) final {
-		return listFiles_impl(path, m_path);
-	}
-
-	Future<Void> deleteContainer(int* pNumDeleted) final {
-		// In order to avoid deleting some random directory due to user error, first describe the backup
-		// and make sure it has something in it.
-		return map(describeBackup(false, invalidVersion), [=](BackupDescription const& desc) {
-			// If the backup has no snapshots and no logs then it's probably not a valid backup
-			if (desc.snapshots.size() == 0 && !desc.minLogBegin.present()) throw backup_invalid_url();
-
-			int count = platform::eraseDirectoryRecursive(m_path);
-			if (pNumDeleted != nullptr) *pNumDeleted = count;
-
-			return Void();
-		});
-	}
-
-private:
-	std::string m_path;
 };
 
 class BackupContainerBlobStore final : public BackupContainerFileSystem, ReferenceCounted<BackupContainerBlobStore> {
